Create new file for zamboni logic (in progress)
